backend/companies/Altis/requirements.txt
python-dotenv
playwright
beautifulsoup4
html2text
openai
requests
